+++ ARController/Sources/Protocols/NodeTappable.swift
@@ -13,24 +13,28 @@ public protocol NodeTappable: NodeContorolling {
 }
 
 public protocol TapGestureRecognisable: class {
-    func processTapGestureAction(_ sender: UITapGestureRecognizer)
+    @discardableResult
+    func processTapGestureAction(_ sender: UITapGestureRecognizer) -> Bool
 }
 
 public extension SceneControlling  where Self: TapGestureRecognisable  {
     
-    public func processTapGestureAction(_ sender: UITapGestureRecognizer) {
-        guard sender.state == .ended else { return }
+    @discardableResult
+    public func processTapGestureAction(_ sender: UITapGestureRecognizer) -> Bool {
+        guard sender.state == .ended else { return false }
         let touchLocation = sender.location(in: sceneView)
-        
         let nodes = self.sceneView.nodesAt(point: touchLocation)
         
+        var gestureProcessed = false
         for each in nodes {
             if let controller: NodeTappable = each.enclosedController() {
                 controller.didTapped(point: touchLocation, node: each, frame: sceneView.session.currentFrame)
+                gestureProcessed = true
                 // Break cycle as soon as some controller able to process the touch
                 break;
             }
         }
+        return gestureProcessed
     }
 }